Implimintin authiring using Get

+++ app/Http/Controllers/QuestionController.php
@@ -53,7 +53,9 @@ class QuestionController extends Controller
      */
     public function show(Question $question)
     {
-        //
+        $question->increment('views');
+
+        return view('question.show', compact('question'));
     }
 
     /**
@@ -63,7 +65,11 @@ class QuestionController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit(Question $question)
-    {
+    {   
+        if(\Gate::denies('update-question', $question)){
+
+            abort(403, "Access denite");
+        }  
         return view("question.edit", compact('question'));
     }
 
@@ -76,6 +82,10 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+
+            abort(403, "Access denite");
+        } 
         $question->update($request->only('title', 'body'));
         
         return redirect('/questions')->with('success', "Your question has been updated.");
@@ -90,10 +100,19 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question)){
+
+            abort(403, "Access denite");
+        } 
         //$this->authorize("delete", $question);
         $question->delete();
 
         return redirect('/questions')->with('success', "Your question has been deleted.");
 
+    }
+
+    private function getBodyHtmlAttribute()
+    {
+        return \Parsedown::instance()->text($this->body);
     }
 }
